Switch to Using Static Library
• Instead of using OneSignal iOS SDK source files and manually adding them to the project, I've switched the framework to instead use a static OneSignal library instead. This should make it easier to update
• Fixed the setLogLevel() function. Calling this function from a JS React-Native file would have crashed the application, switched it to accept integer parameters.
