--- ios/OneSignalNativeSDK/OneSignal/OneSignalReachability.m
@@ -1,171 +1,0 @@
-/**
- * Modified MIT License
- *
- * Copyright 2016 OneSignal
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * 1. The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * 2. All copies of substantial portions of the Software may only be used in connection
- * with services provided by OneSignal.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-#import <arpa/inet.h>
-#import <ifaddrs.h>
-#import <netdb.h>
-#import <sys/socket.h>
-
-#import <CoreFoundation/CoreFoundation.h>
-
-#import "OneSignalReachability.h"
-
-
-#pragma mark - Reachability implementation
-
-@implementation OneSignalReachability
-{
-    BOOL _alwaysReturnLocalWiFiStatus; //default is NO
-    SCNetworkReachabilityRef _reachabilityRef;
-}
-
-+ (instancetype)reachabilityWithAddress:(const struct sockaddr_in *)hostAddress
-{
-    SCNetworkReachabilityRef reachability = SCNetworkReachabilityCreateWithAddress(kCFAllocatorDefault, (const struct sockaddr *)hostAddress);
-    
-    OneSignalReachability* returnValue = NULL;
-    
-    if (reachability != NULL)
-    {
-        returnValue = [[self alloc] init];
-        if (returnValue != NULL)
-        {
-            returnValue->_reachabilityRef = reachability;
-            returnValue->_alwaysReturnLocalWiFiStatus = NO;
-        }
-    }
-    return returnValue;
-}
-
-+ (instancetype)reachabilityForInternetConnection
-{
-    struct sockaddr_in zeroAddress;
-    bzero(&zeroAddress, sizeof(zeroAddress));
-    zeroAddress.sin_len = sizeof(zeroAddress);
-    zeroAddress.sin_family = AF_INET;
-
-    return [self reachabilityWithAddress:&zeroAddress];
-}
-
-- (void)dealloc
-{
-    if (_reachabilityRef != NULL)
-    {
-        CFRelease(_reachabilityRef);
-    }
-}
-
-
-#pragma mark - Network Flag Handling
-
-- (NetworkStatus)localWiFiStatusForFlags:(SCNetworkReachabilityFlags)flags {
-
-    if ((flags & kSCNetworkReachabilityFlagsReachable) && (flags & kSCNetworkReachabilityFlagsIsDirect)) {
-        return ReachableViaWiFi;
-    }
-    return NotReachable;
-}
-
-
-- (NetworkStatus)networkStatusForFlags:(SCNetworkReachabilityFlags)flags {
-    if ((flags & kSCNetworkReachabilityFlagsReachable) == 0) {
-        // The target host is not reachable.
-        return NotReachable;
-    }
-    
-    NetworkStatus returnValue = NotReachable;
-    
-    if ((flags & kSCNetworkReachabilityFlagsConnectionRequired) == 0) {
-        /*
-         If the target host is reachable and no connection is required then we'll assume (for now) that you're on Wi-Fi...
-         */
-        returnValue = ReachableViaWiFi;
-    }
-    
-    if ((((flags & kSCNetworkReachabilityFlagsConnectionOnDemand ) != 0) ||
-         (flags & kSCNetworkReachabilityFlagsConnectionOnTraffic) != 0)) {
-        /*
-         ... and the connection is on-demand (or on-traffic) if the calling application is using the CFSocketStream or higher APIs...
-         */
-        
-        if ((flags & kSCNetworkReachabilityFlagsInterventionRequired) == 0) {
-            /*
-             ... and no [user] intervention is needed...
-             */
-            returnValue = ReachableViaWiFi;
-        }
-    }
-    
-    if ((flags & kSCNetworkReachabilityFlagsIsWWAN) == kSCNetworkReachabilityFlagsIsWWAN)
-    {
-        /*
-         ... but WWAN connections are OK if the calling application is using the CFNetwork APIs.
-         */
-        returnValue = ReachableViaWWAN;
-    }
-    
-    return returnValue;
-}
-
-
-- (BOOL)connectionRequired
-{
-    NSAssert(_reachabilityRef != NULL, @"connectionRequired called with NULL reachabilityRef");
-    SCNetworkReachabilityFlags flags;
-    
-    if (SCNetworkReachabilityGetFlags(_reachabilityRef, &flags))
-    {
-        return (flags & kSCNetworkReachabilityFlagsConnectionRequired);
-    }
-    
-    return NO;
-}
-
-
-- (NetworkStatus)currentReachabilityStatus
-{
-    NSAssert(_reachabilityRef != NULL, @"currentNetworkStatus called with NULL SCNetworkReachabilityRef");
-    NetworkStatus returnValue = NotReachable;
-    SCNetworkReachabilityFlags flags;
-    
-    if (SCNetworkReachabilityGetFlags(_reachabilityRef, &flags))
-    {
-        if (_alwaysReturnLocalWiFiStatus)
-        {
-            returnValue = [self localWiFiStatusForFlags:flags];
-        }
-        else
-        {
-            returnValue = [self networkStatusForFlags:flags];
-        }
-    }
-    
-    return returnValue;
-}
-
-
-@end
